requirements.txt
channels==3.0.4
channels-redis==3.4.0

Django==4.0.4
django-channels==0.7.0
django-environ==0.8.1
django-mptt==0.13.4
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
flake8==4.0.1
phonenumbers==8.12.48
Pillow==9.1.1
psycopg2==2.9.3
redis==4.3.1


